# contract/openapi.yaml
openapi: 3.0.3
info:
  title: AIDetector API
  version: 1.0.0
  description: V1.0 active contract baseline for auth, detection, history, quota, admin, report, key, and team APIs.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /api/v1/history:
    get:
      tags: [history]
      summary: List current user history
      operationId: listHistory
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: History list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [history]
      summary: Create history record
      operationId: createHistory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRecordCreateRequest'
      responses:
        '201':
          description: History record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRecordResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [history]
      summary: Clear all current user history
      operationId: clearHistory
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All history cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearAllResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/history/batch-delete:
    post:
      tags: [history]
      summary: Batch delete history records
      operationId: batchDeleteHistory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDeleteRequest'
      responses:
        '200':
          description: Batch delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeleteResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/history/{historyId}:
    get:
      tags: [history]
      summary: Get history detail
      operationId: getHistory
      security:
        - BearerAuth: []
      parameters:
        - name: historyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: History detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRecordResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: History not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - detail
      properties:
        code:
          oneOf:
            - type: integer
            - type: string
        message:
          type: string
        detail:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string
...
    DetectionFunction:
      type: string
      enum: [scan]
...
    HistorySummary:
      type: object
      required:
        - ai
        - mixed
        - human
      properties:
        ai:
          type: integer
          minimum: 0
          maximum: 100
        mixed:
          type: integer
          minimum: 0
          maximum: 100
        human:
          type: integer
          minimum: 0
          maximum: 100
    HistorySentence:
      type: object
      required:
        - id
        - text
        - raw
        - type
        - probability
        - score
        - reason
        - suggestion
      properties:
        id:
          type: string
        text:
          type: string
        raw:
          type: string
        startParagraph:
          type: integer
          nullable: true
          minimum: 1
        endParagraph:
          type: integer
          nullable: true
          minimum: 1
        type:
          type: string
          enum: [ai, mixed, human]
        probability:
          type: number
          format: float
          minimum: 0
          maximum: 1
        score:
          type: integer
          minimum: 0
          maximum: 100
        reason:
          type: string
        suggestion:
          type: string
    HistoryCitation:
      type: object
      required:
        - id
        - text
        - source
      properties:
        id:
          type: string
        text:
          type: string
        source:
          type: string
    HistoryAnalysis:
      type: object
      description: >
        Analysis payload used by current scan/history flows. The translation, polish, and citations
        fields are compatibility placeholders and are not populated by active product features.
      required:
        - summary
        - sentences
        - translation
        - polish
        - citations
        - aiLikelyCount
        - highlightedHtml
      properties:
        summary:
          $ref: '#/components/schemas/HistorySummary'
        sentences:
          type: array
          items:
            $ref: '#/components/schemas/HistorySentence'
        translation:
          type: string
          description: Reserved compatibility field. Currently returned as an empty string.
        polish:
          type: string
          description: Reserved compatibility field. Currently returned as an empty string.
        citations:
          type: array
          description: Reserved compatibility field. Currently returned as an empty array.
          items:
            $ref: '#/components/schemas/HistoryCitation'
        aiLikelyCount:
          type: integer
          minimum: 0
        highlightedHtml:
          type: string
    HistoryRecordCreateRequest:
      type: object
      required:
        - title
        - functions
        - inputText
        - editorHtml
      properties:
        title:
          type: string
          maxLength: 255
        functions:
          type: array
          items:
            $ref: '#/components/schemas/DetectionFunction'
        inputText:
          type: string
          maxLength: 50000
        editorHtml:
          type: string
        analysis:
          $ref: '#/components/schemas/HistoryAnalysis'
          nullable: true
...
    HistoryRecordResponse:
      type: object
      required:
        - id
        - userId
        - createdAt
        - functions
        - inputText
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        functions:
          type: array
          items:
            $ref: '#/components/schemas/DetectionFunction'
        inputText:
          type: string
        editorHtml:
          type: string
          nullable: true
        analysis:
          $ref: '#/components/schemas/HistoryAnalysis'
          nullable: true
    HistoryListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryRecordResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    BatchDeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: integer
    BatchDeleteResponse:
      type: object
      required:
        - deletedCount
        - failedIds
      properties:
        deletedCount:
          type: integer
        failedIds:
          type: array
          items:
            type: integer
    ClearAllResponse:
      type: object
      required:
        - deletedCount
      properties:
        deletedCount:
          type: integer
...